refactor(background): split background code into a library
Move background job management, dialog, and key bindings into
separate modules under background/ (libbackground).
